requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
python-docx>=1.1.0
mammoth>=1.6.0
weasyprint>=60.0
docx2pdf>=0.1.8
